Exclude DNT by default: According to this site https://stackoverflow.com/questions/32890865/does-private-mode-enable-do-not-track-dnt-on-popular-browsers most browsers enable DNT in incognito mode. I can confirm it is automatically enabled in latest Firefox and Safari, but not in latest Chrome.
I guess it makes no sense to consider DNT for these browsers in the first place. If we're in incognito mode, there's no way of knowing if the user has DNT enabled in non-incognito mode too, which means we cannot know if the DNT component tampers with the fingerprint. So while this removes some of the entropy, it makes fingerprinting robust against incognito mode.